requirements.txt
fastapi
uvicorn
transformers
redis
rarfile
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
